# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_jhoicas_Inventario-api_internal_application_dto.EmailAccountListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.EmailAccountResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.EmailAccountResponse:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      email_address:
        type: string
      id:
        type: string
      imap_port:
        type: integer
      imap_server:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.EmailAttachmentResponse:
    properties:
      file_name:
        type: string
      file_url:
        type: string
      id:
        type: string
      mime_type:
        type: string
      size:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.EmailListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.EmailResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.EmailResponse:
    properties:
      account_id:
        type: string
      attachments:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.EmailAttachmentResponse'
        type: array
      body_html:
        type: string
      body_text:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      from_address:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message_id:
        type: string
      received_at:
        type: string
      subject:
        type: string
      to_address:
        type: string
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
host: api.naturerp.ludoia.com
info:
  contact: {}
  description: API para el sistema ERP.
  title: Tu API ERP
  version: "1.0"
paths:
...
  /api/emails:
    get:
      description: 'Respuesta paginada estandar: {items,total,limit,offset}'
      parameters:
      - description: Filtrar por cliente
        in: query
        name: customer_id
        type: string
      - description: Filtrar por leidos/no leidos
        in: query
        name: is_read
        type: boolean
      - default: 20
        description: Limite
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.EmailListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Listar emails
      tags:
      - email
...
  /api/settings/email-accounts:
    get:
      description: 'Respuesta paginada estandar: {items,total,limit,offset}'
      parameters:
      - default: 20
        description: Limite
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.EmailAccountListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Listar cuentas de email
      tags:
      - email
...
securityDefinitions:
  Bearer:
    description: |-
      Escribe "Bearer " seguido de un espacio y tu token.
      Escribe "Bearer " seguido de un espacio y tu token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"